@@ -4,14 +4,14 @@ description: WidgetFactory extension to render VIDEO with the chewie plugin.
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=2.0.0"
+  flutter: ">=2.2.0"
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   chewie: ^1.0.0
   flutter:
     sdk: flutter
-  flutter_widget_from_html_core: ">=0.6.1 <0.8.0"
+  flutter_widget_from_html_core: ^0.7.0-0
   video_player: ^2.0.0
 
 dependency_overrides:
