@@ -42,13 +42,14 @@ class HtmlWidget extends core.HtmlWidget {
   HtmlWidget(
     String html, {
     bool? buildAsync,
-    AsyncWidgetBuilder<Widget>? buildAsyncBuilder,
     bool enableCaching = true,
     WidgetFactory Function()? factoryBuilder,
     Key? key,
     Uri? baseUrl,
     CustomStylesBuilder? customStylesBuilder,
     CustomWidgetBuilder? customWidgetBuilder,
+    OnErrorBuilder? onErrorBuilder,
+    OnLoadingBuilder? onLoadingBuilder,
     void Function(ImageMetadata)? onTapImage,
     FutureOr<bool> Function(String)? onTapUrl,
     core.RebuildTriggers? rebuildTriggers,
@@ -63,11 +64,12 @@ class HtmlWidget extends core.HtmlWidget {
           html,
           baseUrl: baseUrl,
           buildAsync: buildAsync,
-          buildAsyncBuilder: buildAsyncBuilder,
           customStylesBuilder: customStylesBuilder,
           customWidgetBuilder: customWidgetBuilder,
           enableCaching: enableCaching,
           factoryBuilder: factoryBuilder ?? _getEnhancedWf,
+          onErrorBuilder: onErrorBuilder,
+          onLoadingBuilder: onLoadingBuilder,
           onTapImage: onTapImage,
           onTapUrl: onTapUrl,
           rebuildTriggers: core.RebuildTriggers([
