@@ -18,6 +18,8 @@ final hwKey = helper.hwKey;
 
 const buildCurrentState = helper.buildCurrentState;
 
+const explainWithoutPumping = helper.explainWithoutPumping;
+
 String? _explainer(helper.Explainer parent, Widget widget) {
   return fwfh_cached_network_image.cachedNetworkImageExplainer(
           parent, widget) ??
@@ -30,14 +32,12 @@ String? _explainer(helper.Explainer parent, Widget widget) {
 Future<String> explain(
   WidgetTester tester,
   String? html, {
-  bool buildFutureBuilderWithData = true,
   Widget? hw,
   bool useExplainer = true,
 }) async =>
     helper.explain(
       tester,
       null,
-      buildFutureBuilderWithData: buildFutureBuilderWithData,
       explainer: _explainer,
       hw: hw ?? HtmlWidget(html!, key: helper.hwKey),
       useExplainer: useExplainer,
